Validations on selecting slot and date and booking slot for logged in user
Validations to ,
check if the selected date has passed. If passed user wont be able to book.
check if the selected slot is already booked
check if the selected slot is an invalid number

Book the slot for current logged in user name instead of hardcoded value
In DB store the username + emailID format for uniqueness and in UI show only user name.